Fix duplicated body tag
The css class is needed on body tag for bulma to use correct spacing under navbar.
Helmet can be used to add the tag properly, it could be done inside the
navbar component but there's a risk of it interfering later - if
something else needs to be added to body.